perf(cli): fast-path project start
